requirements.txt
streamlit
rasterio
numpy
tensorflow
pillow
matplotlib